pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "chasten", email = "literallywar@gmail.com"},
]
dependencies = [
    "requests>=2.28.2",
    "numpy>=1.24.2",
    "tqdm>=4.65.0",
    "openai>=0.27.1",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
